mcp_local_server_agent/agent/requirements.txt
azure-ai-projects
azure-ai-agents>=1_1.2.0b2
azure-identity
python-dotenv>=1.0